Moved it all to parser package
